Move and rename user types entity access classes
