+++ web/modules/user_types/creatives/src/Access/CreativeEntityAccess.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\creatives;
+namespace Drupal\creatives\Access;
 
 use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Entity\EntityInterface;
@@ -10,7 +10,7 @@ use Drupal\creatives\Entity\Creative;
 /**
  * Access controller for the Creative entity.
  */
-class CreativeAccessControlHandler {
+class CreativeEntityAccess {
 
   /**
    * Checks access.
